10/24/12 game: add an outline around face, because we have a bright yellow face on a light yellow background 10/24/12 game: reduce size of face, change layout so that face is below equation boxes 10/24/12 game: reduce equation font size so that everything fits in equation boxes

+++ simulations-java/simulations/line-graphing/src/edu/colorado/phet/linegraphing/linegame/model/maketheequation/MTE_Challenge.java
@@ -23,7 +23,7 @@ import edu.colorado.phet.linegraphing.linegame.model.Challenge;
 public abstract class MTE_Challenge extends Challenge {
 
     private static final int GRAPH_WIDTH = 400; // graph width in view coordinates
-    private static final Point2D ORIGIN_OFFSET = new Point2D.Double( 300, 300 ); // offset of the origin (center of the graph) in view coordinates
+    private static final Point2D ORIGIN_OFFSET = new Point2D.Double( 275, 300 ); // offset of the origin (center of the graph) in view coordinates
 
     public final ModelViewTransform mvt; // transform between model and view coordinate frames
     public final Graph graph; // the graph that plots the lines
